Refactor MovieCard div outside of hero element.
